# client-spec/openapi/auth-api.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Authentication and authorization service for jorisjonkers.dev",
    "title" : "Auth API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://auth.jorisjonkers.dev",
    "description" : "Production"
  }, {
    "url" : "http://localhost:8081",
    "description" : "Local development"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/api/v1/admin/users/{id}" : {
      "delete" : {
        "operationId" : "deleteUser",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : { },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-controller" ]
      },
      "get" : {
        "operationId" : "getUser",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminUserResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-controller" ]
      }
    },
    "/api/v1/admin/users/{id}/role" : {
      "patch" : {
        "operationId" : "updateRole",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateRoleRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminUserResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-controller" ]
      }
    },
    "/api/v1/admin/users/{id}/services" : {
      "put" : {
        "operationId" : "updateServicePermissions",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateServicePermissionsRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminUserResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-controller" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "AdminUserResponse" : {
        "type" : "object",
        "properties" : {
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "email" : {
            "type" : "string"
          },
          "emailConfirmed" : {
            "type" : "boolean"
          },
          "firstName" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "lastName" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          },
          "servicePermissions" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "totpEnabled" : {
            "type" : "boolean"
          },
          "username" : {
            "type" : "string"
          }
        },
        "required" : [ "createdAt", "email", "emailConfirmed", "firstName", "id", "lastName", "role", "servicePermissions", "totpEnabled", "username" ]
      },
...
      "UpdateRoleRequest" : {
        "type" : "object",
        "properties" : {
          "role" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "role" ]
      },
      "UpdateServicePermissionsRequest" : {
        "type" : "object",
        "properties" : {
          "services" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "services" ]
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      },
      "oauth2" : {
        "flows" : {
          "authorizationCode" : {
            "authorizationUrl" : "https://auth.jorisjonkers.dev/api/oauth2/authorize",
            "refreshUrl" : "https://auth.jorisjonkers.dev/api/oauth2/token",
            "tokenUrl" : "https://auth.jorisjonkers.dev/api/oauth2/token"
          }
        },
        "type" : "oauth2"
      }
    }
  }
}